Added kube-conformance service
This patch adds the kube-conformance service to the first non-worker
node installed in a Kubernetes cluster. The service is not started
automatically. Starting the service downloads the version of
kubernetes-test.tar.gz that matches the installed version of Kubernetes
on the cluster. The service then runs the conformance test suite by
invoking /opt/bin/run-kube-conformance.sh.